refactor(workspace): introduce leo-cli-core for shared CLI/wasm core
The wasm-buildable surface that used to live inside `crates/leo`
(forcing the CLI crate into a dual binary+lib role with a partially
gated dependency tree) and `crates/leo-wasm`'s `project.rs` now live
in a single new library crate: `leo-cli-core`.

Layout:

  crates/leo-cli-core/        ← shared, wasm-buildable
    src/options.rs            EnvOptions, BuildOptions, From<BuildOptions>
                              for CompilerOptions
    src/project.rs            Project, compile/compile_test,
                              stage_programs, run_function,
                              find_test_functions, TestFn

  crates/leo/                 ← CLI binary, depends on leo-cli-core
    src/cli/commands/common/options.rs
      `pub use leo_cli_core::options::{BuildOptions, EnvOptions,
       DEFAULT_ENDPOINT};` keeps every existing CLI import working.

  crates/leo-wasm/            ← wasm-bindgen layer, depends on leo-cli-core
    src/wire.rs               `pub use leo_cli_core::options::*;`
    src/commands/*.rs         call `leo_cli_core::project::*` directly

Net effect: both consumers are now genuine thin wrappers around one
shared core. `crates/leo`'s lib drops the wasm-buildable carve-out
(the cfg-gated `pub mod cli;` and the bespoke top-level `options` /
`project` modules) — the lib is straight-up native again.
`crates/leo-wasm`'s `project.rs` is deleted entirely; the file map
loader, dep-graph walker, and execution helpers live in one place.

Follow-up: lift `handle_build` / `handle_run` / `handle_test` (and
their FileSink abstraction) into `leo-cli-core` so the CLI's
clap-driven structs and `leo-wasm`'s `*_impl` entry points both
become trivial wrappers around `leo_cli_core::handle_*`.
